% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{corrupt_graph}
\alias{corrupt_graph}
\title{Corrupt a graph by random edge swaps}
\usage{
corrupt_graph(A, corruption_rate = 0.3)
}
\arguments{
\item{A}{Symmetric p x p binary adjacency matrix.}

\item{corruption_rate}{Proportion of edges to remove and replace.}
}
\value{
A corrupted adjacency matrix.
}
\description{
Corrupt a graph by random edge swaps
}
