% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{generate_data}
\alias{generate_data}
\title{Simulate data for graph-IV regression}
\usage{
generate_data(
  n = 100,
  p = 70,
  q = 500,
  s1 = 0.1,
  s_alpha = 10,
  alpha_strength = 5,
  beta_true
)
}
\arguments{
\item{n}{Sample size.}

\item{p}{Number of exposures.}

\item{q}{Number of instruments.}

\item{s1}{Fraction of instruments relevant for each exposure.}

\item{s_alpha}{Number of invalid instruments.}

\item{alpha_strength}{Direct-effect magnitude of invalid instruments.}

\item{beta_true}{Numeric vector of length p of true causal effects.}
}
\value{
A list with \code{Y}, \code{X}, \code{Z}, \code{A_true},
and \code{alpha_true}.
}
\description{
Simulate data for graph-IV regression
}
