% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_laplacian}
\alias{get_laplacian}
\title{Compute the unnormalised graph Laplacian}
\usage{
get_laplacian(A)
}
\arguments{
\item{A}{Symmetric p x p binary adjacency matrix.}
}
\value{
A p x p Laplacian matrix.
}
\description{
Compute the unnormalised graph Laplacian
}
