% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mcc}
\alias{get_mcc}
\title{Matthews Correlation Coefficient for support recovery}
\usage{
get_mcc(true_support, estimated_support, p)
}
\arguments{
\item{true_support}{Integer vector of truly active indices.}

\item{estimated_support}{Integer vector of estimated active indices.}

\item{p}{Total number of predictors.}
}
\value{
A scalar between -1 and 1.
}
\description{
Matthews Correlation Coefficient for support recovery
}
