% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{iv_lasso}
\alias{iv_lasso}
\title{IV-LASSO: Two-stage LASSO without graph structure}
\usage{
iv_lasso(Y, X, Z)
}
\arguments{
\item{Y}{Numeric vector of length n. Outcome.}

\item{X}{Numeric n x p matrix of endogenous exposures.}

\item{Z}{Numeric n x q matrix of instruments.}
}
\value{
Numeric vector of length p of estimated causal effects.
}
\description{
IV-LASSO: Two-stage LASSO without graph structure
}
