% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{ivgl}
\alias{ivgl}
\title{IVGL: IV regression with graph-fused Lasso}
\usage{
ivgl(Y, X, Z, L)
}
\arguments{
\item{Y}{Numeric vector of length n. Outcome.}

\item{X}{Numeric n x p matrix of endogenous exposures.}

\item{Z}{Numeric n x q matrix of instruments.}

\item{L}{Numeric p x p graph Laplacian (see \code{\link{get_laplacian}}).}
}
\value{
Numeric vector of length p of estimated causal effects.
}
\description{
IVGL: IV regression with graph-fused Lasso
}
